fix deprecation

+++ utils/modules/matrix_utils.py
@@ -52,6 +52,6 @@ def matrix_apply_np(verts, matrix):
     verts should be a numpy array with shape (n,3)
     matrix can be a regular mathultis matrix'''
 
-    verts_co_4d = np.ones(shape=(verts.shape[0], 4), dtype=np.float)
+    verts_co_4d = np.ones(shape=(verts.shape[0], 4), dtype=np.float32)
     verts_co_4d[:, :-1] = verts  # cos v (x,y,z,1) - point,   v(x,y,z,0)- vector
     return np.einsum('ij,aj->ai', matrix, verts_co_4d)[:, :-1]
